+++ app/routes/components/Background/changelog.codegen.ts
@@ -4,4 +4,5 @@
 export const changelog = ''
   + '- 2025-09-30: Initial version.\n'
   + '- 2025-10-04: Update code style.\n'
+  + '- 2025-10-25: Support `style` prop and add `Background.Container` sub-component.\n'
   ;