+++ app/routes/components/Background/index.tsx
@@ -1,5 +1,6 @@
-import { ComponentTemplate } from "@/components/ComponentTemplate";
-import { Section } from "@/components/Section";
+import { useState } from "react";
+
+import { DocLayout } from "@/layouts/DocLayout/doc-layout";
 
 import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
@@ -15,13 +16,42 @@ export function meta(_: Route.MetaArgs) {
 }
 
 export default function BackgroundPage() {
+  const [tabsValue, setTabsValue] = useState("doc");
+
+  const tabs = [
+    { name: "doc", label: "Document" },
+    { name: "source", label: "Source Code" },
+    { name: "changelog", label: "Changelog" },
+  ];
+
   return (
-    <ComponentTemplate component="Background">
-      <Section.Root title="Background">
-        <Section.Demo node={<Demo01 />} code={demo01SourceCode} />
-        <Section.SourceCode files={sourceCode} />
-        <Section.Changelog changelog={changelog} />
-      </Section.Root>
-    </ComponentTemplate>
+    <DocLayout>
+      <DocLayout.Title
+        title="Background"
+        desc="More backgrounds you want."
+        tabsValue={tabsValue}
+        tabs={tabs}
+        onTabsValueChange={setTabsValue}
+      />
+      {tabsValue === "doc" && (
+        <DocLayout.Content>
+          <DocLayout.Live
+            nodeBgKind="none"
+            node={<Demo01 />}
+            code={demo01SourceCode}
+          />
+        </DocLayout.Content>
+      )}
+      {tabsValue === "source" && (
+        <DocLayout.Content>
+          <DocLayout.Files files={sourceCode} />
+        </DocLayout.Content>
+      )}
+      {tabsValue === "changelog" && (
+        <DocLayout.Content>
+          <DocLayout.Changelog changelog={changelog} />
+        </DocLayout.Content>
+      )}
+    </DocLayout>
   );
 }